Handle scoring documents with "score" attributes.
While scoring document parts, we stuff values in "score" attributes of nodes.  Then we compare various scores.  Our scores are floats, but if the document has a `score="foo"` attribute in it, that will be a string.  Inequality between float and string does not work, so only use float values -- which must have come from us.